+++ packages/core/src/utils/createMap/defaults.ts
@@ -30,7 +30,7 @@ export default {
     ],
     [
       'EPSG:31467',
-      '+proj=tmerc +lat_0=0 +lon_0=9 +k=1 +x_0=3500000 +y_0=0 +ellps=bessel +nadgrids=BETA2007.gsb +units=m +no_defs +type=crs',
+      '+proj=tmerc +lat_0=0 +lon_0=9 +k=1 +x_0=3500000 +y_0=0 +ellps=bessel +towgs84=598.1,73.7,418.2,0.202,0.045,-2.455,6.70 +units=m +no_defs',
     ],
     [
       'EPSG:4647',