+++ packages/core/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## unpublished
+
+- Fix: Change `EPSG:31467` projection definition string to be usable for `proj` conversions.
+
 ## 3.2.0
 
 - Feature: Default core export now contains a `resetPlugins` method that allows undoing all previous `addPlugins` calls.